test: 당첨 번호, 보너스 번호 테스트 코드 분리
